+++ fapi/src/main/java/by/training/nc/sd3/service/UserAccountDataService.java
@@ -5,7 +5,7 @@ import by.training.nc.sd3.models.UserAccountViewModel;
 import java.util.List;
 
 public interface UserAccountDataService {
-    UserAccountViewModel getUserAccountById(int id);
+    UserAccountViewModel getUserAccountById(Long id);
     UserAccountViewModel getUserAccountByData(String login, String password);
     UserAccountViewModel saveUserAccount(UserAccountViewModel account);
 }
